refactor file stasher, response stasher
