+++ packages/core/package.json
@@ -52,6 +52,7 @@
     "node-fetch": "2.6.7",
     "oauth-sign": "0.9.0",
     "semver": "7.5.2",
+    "uuid": "^10.0.0",
     "zapier-platform-schema": "15.8.0"
   },
   "devDependencies": {
