+++ packages/core/src/app-middlewares/after/large-response-cacher.js
@@ -2,7 +2,7 @@
 
 const constants = require('../../constants');
 const cleaner = require('../../tools/cleaner');
-const createResponseStasher = require('../../tools/create-response-stasher');
+const responseStasher = require('../../tools/create-response-stasher');
 
 /*
   TODO: Some services _do not_ enjoy having 6mb+ responses returned, so
@@ -14,16 +14,16 @@ const largeResponseCachePointer = async (output) => {
 
   const autostashOutputLimit =
     output.input._zapier.event.autostashPayloadOutputLimit;
-  const size = JSON.stringify(response).length;
+
+  const payload = JSON.stringify(response.results);
+  const size = payload.length;
   if (size > constants.RESPONSE_SIZE_LIMIT) {
     console.log(
       `Oh no! Payload is ${size}, which is larger than ${constants.RESPONSE_SIZE_LIMIT}.`
     );
 
     if (autostashOutputLimit && size <= autostashOutputLimit) {
-      const stasher = createResponseStasher(output.input);
-
-      const url = await stasher(JSON.stringify(output.results));
+      const url = await responseStasher(output.input, payload, size);
       output.resultsUrl = url;
       output.results = Array.isArray(output.results) ? [] : {};
     }
